画面構成 HTMLの編集

+++ src/main/java/com/example/Todo/entity/UserInf.java
@@ -2,7 +2,7 @@ package com.example.Todo.entity;
 
 public interface UserInf {
 
-	Long getUserId();
+	int getUserId();
 
 	String getUsername();
 }
